+++ src/export/portfolio_helpers.py
@@ -1,7 +1,7 @@
 # src/export/portfolio_helpers.py
 from __future__ import annotations
 
-from typing import Any, List
+from typing import Any, List, Dict
 
 from .shared_helpers import (
     parse_date,
@@ -10,9 +10,9 @@ from .shared_helpers import (
     clean_languages_above_threshold,
 )
 
-# -------------------------
+from src.insights.portfolio import format_skills_block, format_languages, format_frameworks
+
 # Portfolio-only helpers
-# -------------------------
 
 def reformat_duration_line(duration_line: str) -> str:
     """
@@ -45,13 +45,54 @@ def reformat_duration_line(duration_line: str) -> str:
     d = parse_date(s)
     return d.strftime("%b %Y") if d else ""
 
+def _skills_one_line(summary: Dict[str, Any]) -> str:
+    lines = format_skills_block(summary) or []
+    if not lines:
+        return ""
+    if len(lines) == 1 and "N/A" in (lines[0] or ""):
+        return ""
+
+    out: List[str] = []
+    for line in lines:
+        t = (line or "").strip()
+        if not t:
+            continue
+        # skip "Skills:" header if present
+        if t.lower().startswith("skills"):
+            continue
+        if t.startswith(("-", "•")):
+            t = t.lstrip("-•").strip()
+        if t:
+            out.append(t)
+    return ", ".join(out)
+
+def _languages_clean(summary: Dict[str, Any], *, min_pct: int = 10) -> str:
+    # Prefer raw list with percents -> threshold + remove percent
+    langs = clean_languages_above_threshold(summary.get("languages"), min_pct=min_pct)
+    if langs:
+        return ", ".join(langs)
+
+    # Fallback: formatted string -> strip percent tokens
+    s = strip_percent_tokens(format_languages(summary) or "")
+    if s.lower().startswith("languages:"):
+        s = s.split(":", 1)[1].strip()
+    return s.strip() or "N/A"
+
+def _frameworks_clean(summary: Dict[str, Any]) -> str:
+    s = strip_percent_tokens(format_frameworks(summary) or "")
+    if s.lower().startswith("frameworks:"):
+        s = s.split(":", 1)[1].strip()
+    return s.strip() or "N/A"
 
 __all__ = [
-    # re-export shared helpers (optional convenience)
+    # re-export shared helpers
     "parse_date",
     "format_date_range",
     "strip_percent_tokens",
     "clean_languages_above_threshold",
     # portfolio-only
     "reformat_duration_line",
+    "_skills_one_line",
+    "_languages_clean",
+    "_frameworks_clean"
 ]
