+++ src/export/portfolio_pdf.py
@@ -40,24 +40,19 @@ from src.insights.rank_projects.rank_project_importance import collect_project_d
 from src.db import get_project_summary_row, get_project_thumbnail_path
 from src.insights.portfolio import (
     format_duration,
-    format_languages,
-    format_frameworks,
     format_activity_line,
-    format_skills_block,
     resolve_portfolio_display_name,
     resolve_portfolio_summary_text,
     resolve_portfolio_contribution_bullets,
 )
 
-# ✅ use shared helpers (single source of truth)
+# use shared helpers
 from src.export.shared_helpers import (
-    parse_date,
     format_date_range,
     strip_percent_tokens,
-    clean_languages_above_threshold,
 )
-from src.export.portfolio_helpers import reformat_duration_line
-
+from src.export.portfolio_helpers import reformat_duration_line, _skills_one_line,  _frameworks_clean, _languages_clean
+from src.insights.portfolio.formatters import _clean_bullets
 
 # -------------------------
 # Local helpers (PDF-only)
@@ -95,64 +90,6 @@ def _load_image_preserve_aspect(path: str, max_width: float) -> Image | None:
         return flow
     except Exception:
         return None
-
-
-def _skills_one_line(summary: Dict[str, Any]) -> str:
-    lines = format_skills_block(summary) or []
-    if not lines:
-        return ""
-    if len(lines) == 1 and "N/A" in (lines[0] or ""):
-        return ""
-
-    out: List[str] = []
-    for line in lines:
-        t = (line or "").strip()
-        if not t:
-            continue
-        # skip "Skills:" header line if present
-        if t.lower().startswith("skills"):
-            continue
-        if t.startswith(("-", "•")):
-            t = t.lstrip("-•").strip()
-        if t:
-            out.append(t)
-    return ", ".join(out)
-
-
-def _languages_clean(summary: Dict[str, Any], *, min_pct: int = 10) -> str:
-    # Prefer raw list with percents -> threshold + remove percent
-    langs = clean_languages_above_threshold(summary.get("languages"), min_pct=min_pct)
-    if langs:
-        return ", ".join(langs)
-
-    # Fallback: formatted string -> strip percent tokens
-    s = strip_percent_tokens(format_languages(summary) or "")
-    if s.lower().startswith("languages:"):
-        s = s.split(":", 1)[1].strip()
-    return s.strip() or "N/A"
-
-
-def _frameworks_clean(summary: Dict[str, Any]) -> str:
-    s = strip_percent_tokens(format_frameworks(summary) or "")
-    if s.lower().startswith("frameworks:"):
-        s = s.split(":", 1)[1].strip()
-    return s.strip() or "N/A"
-
-
-def _clean_bullets(values: Any) -> List[str]:
-    if not isinstance(values, list):
-        return []
-    out: List[str] = []
-    for item in values:
-        t = str(item).strip()
-        if not t:
-            continue
-        if t.startswith(("-", "•")):
-            t = t.lstrip("-•").strip()
-        if t:
-            out.append(t)
-    return out
-
 
 # -------------------------
 # Export
